Fixed sprinkler animation; added cell grid; changed cell to variable size (untested)

+++ src/main/java/lu/kbra/plant_game/engine/entity/go/mesh/terrain/TerrainMesh.java
@@ -72,7 +72,7 @@ public class TerrainMesh extends LoadedMesh {
 		return this.materialType;
 	}
 
-	public int getCellSize() {
+	public float getCellSize() {
 		return 1;
 	}
